freshtools/requirements.txt
peewee==2.10.2
python-dateutil==2.6.1
six==1.11.0

# Development
pytest==3.4.1